Pair of vrt tools to move metadata out and back
Put a vrt document through a "flat" form that can be processed
with tools that do not expect markup between tokens. Markup
storage column is named "+" which is otherwise not a valid
name, and markup is stored with angle brackets as such, which
is not valid vrt. (Er, is this the place for documentation?)

Composition (vrt-flat | vrt-deep) is idempotent on vrt.
Other composition (vrt-deep | vrt-flat) flattens new markup.

scripts/prevrt/vrt-flat
    stores markup and comments inside sentence in an
    annotation column aka in flat form so that tokens
    form a consecutive sequence of lines (some reordering
    also occurs on boundaries, and loss of strangely placed
    positional-attribute comments)

scripts/prevrt/vrt-deep
    returns flatly stored markup and comments to
    their proper positions (apart from the minor
    rearrangements that are not reversible)