Fixed the fillup bug
Was a previously unhandled fallthrough that was never meant to be a
crash. Now handled properly. (Another improper fallthrough remains,
but no test cases are not yet known, and some overrides of Punct, Num,
Prop from the original script are not yet imitated.)

Also made fillup use the new libraries.

vrt-tdp-alpha-fillup
(vrtargslib)
(vrtdatalib)