Worked parser to terminate even in "exceptional" situations
Exceptional situations include broken pipe and keyboard interruption
but also programming errors in vrt-tdp-alpha-parse itself. Actual
parser did not seem to terminate on a broken pipe (was not written to
write its output to stdout anyway) but it did report it when it was
allowed to get it (presumably in its stderr but this needs checked).

(SIGPIPE is probably all right now that the actual parser is not
getting it, but watching its diagnostics might provide a way to kill
it when it runs out of memory - to be investigated.)

vrt-tdp-alpha-parse