vrtargslib.py: accept --backup without --in-place
Backup was already documented in the help text to also have the effect
of in-place but an obsolete check had been forgotten in place when the
options were made mutually exclusive, and various filename operations
needed to depend on both inplace and backup.

Found and fixed using vrt-name but bug and fix affect other tools that
use vrtargslib.