New tool to project vrt to other than undesired fields
In particular, all fields with a dot in their names can be dropped.

vrt-drop

(Warning: like vrt-keep, vrt-drop is desired to keep any '+' field but
this has not been tested yet and is likely to need some fix or other.)

(Also, previous commit of vrt-keep calls fields "arguments". That
makes no sense whatsoever. Sorry about that.)