VRT tools to mend previously torn sentences
One tool to simply remove sentence boundaries between shreds.
One tool to also adjust dependency annotation, assuming CoNLL'09.

(The first was written when it was found that not only the parser but
also the marmot has trouble with excessively long "sentences" - there
was no dependency annotation to fix after the marmot. The second
supersedes vrt-tdp-alpha-mend, to be retired.)

vrt-simple-mend
vrt-conll09-mend