Wrote a batch tool for array jobs
Produces an array job with a separate task for each argument following
the // marker on the command line.

Output file options need work, but otherwise the tool has already been
used with good effect. The biggest concern is that now there are again
two programs (game and gamarray) that share much code, which should be
made a library.

gamarray