[VRT tools] Made marmot tagger tool warn of long "sentences"
Some overlong sentences appeared to cause trouble to marmot to
the extent that they could not be tagged without tearing first.
Now there will be a message in stderr when such are shipped to
marmot. Pay no attention to the warning if marmot finishes -
the warning limit is probably low.

vrt-tdp-alpha-marmot