Have vrt-sparv-cstlemma choose a lemma
Use the simplest imaginable choice that has anything to do with the
available candidates: first candidate from dictionary or else from
rules (or _ just in case it ever happens that even rules give none).

Not sure what cstlemma produces for a vertical bar in a lemma (must
test but even then will not be able to be sure that it does it always
so meh) but normalizing any vertical bar to a broken bar in a lemma.