Handle some abused character codes
Map a couple more C1 codes to what they would be in Windows 1252,
mostly remove some other control codes, remove private use codes,
leave as is a code that was reported as unassigned because newer
(assigned in Unicode 7).

In the end the actions are governed by garbage in, garbage out. Just
making sure the final garbage is plaintext in UTF-8, ISWIM.

libylespecial.py