Refactor tdp pipeline: move a lexical lookup table into a library module
Additional words that override Omorfi are used by two tools. The table
used to be quite minimal but is now about to be expanded with some
nonstandard words found in the Iijoki corpus. It's an experiment. (It
would be better to upgrade the lexical transducer itself.)

new: libtdpalphalookup.py
modified:   vrt-tdp-alpha-fillup
modified:   vrt-tdp-alpha-lookup