scripts: Many Python 2 scripts: python -> python2
Changed "#! /usr/bin/env python" to "... python2" in actively used
Python 2 scripts to make them work in environments where "python" refers
to Python 3, such as CSC's Puhti with the kieli module loaded.

This means that the scripts will no longer work unmodified in
environments with no "python2", but that should not be a major issued,
as one could also write a "python2" script calling "python".

Note that this only affects scripts actively used at present, such as
those called in korp-make. The other scripts should be also modified
similarly if they are to be actively used.