Work-around a bug in some input data: fix even a missing lang
Some KLK files have empty sentences that do not obtain the lang
attribute in the first-pass language identification so those sentences
(that are not really sentences, or anything, at all) do not have the
attribute in the post-pass. Make vrt-fix-text-lang do something fake
insted of crashing.

The real fix is to not have such sentence elements in the first place.

modified:   libvrt/tools/vrt_fix_text_lang.py