Fix language identification post-processor
The meaning was to keep language codes that remain in some sentence in
the text but that final step was never implemented. Is now. (The
processor replaces dubious identifications with xxx, but a remaining
non-xxx indicates that other sentences in that same non-xxx may not
have been so dubious after all.)

modified:   libvrt/tools/vrt_fix_text_lang.py