Actual blocking rules - learned along the way
The logic of skipping some sentences (with _ as output)
was copied from some older tool but the actual patterns
were learned little by little when encountering difficult
sentences while annotating a corpus.

Mainly sequences of cap-initial or all-caps or even a cap
somewhere, and sequences of tokens that end with digits,
allowing various punctuation marks in the tokens.