Adopt a reservoir sampler with intention to use it to sample VRT
This just adds a single libvrt file, not used for anything yet.

The implementation is Copyright 2021 Oscar Benjamin, full license to
do most anything with it is stated on top of the file. See the file.
(This was originally proposed for inclusion in Python but Python did
not want it, see https://bugs.python.org/issue41311, from the file.)