First version of vrt-fix-no-space, the fixer of missing spaces in tokens
Been complaining of this fashion long enough and now new discussion
forum data is imminent, so time to do something about it. The first
tentative pattern is to split tokens like "takia.Niin" in three parts
in two sentences:

takia
.
</sentence>
<sentence>
Niin

There are marking options to inspect the result. More to follow.